+++ esm/index.js
@@ -12,3 +12,4 @@ export { default as uttoriInline } from '../src/uttori-inline';
 export { default as wikilinks } from '../src/wikilinks';
 export { default as youtube } from '../src/youtube';
 export { default as lineBreaker } from '../src/line-breaker';
+export { default as video } from '../src/video';